add checker

+++ src/main/java/oj/judge/common/Problem.java
@@ -3,8 +3,8 @@ package oj.judge.common;
 public class Problem {
     public Long id;
     
-    public int timeLimit; // in ms. 0 for not specified.
-    public int memoryLimit; // in MB. 0 for not specified.
+    public double timeLimit; // in ms. 0 for not specified.
+    public double memoryLimit; // in MB. 0 for not specified.
     public boolean specialJudge;
 
     public String resourcesHash;
